server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==0.28.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
google-api-python-client==2.108.0

# Database dependencies
sqlalchemy==1.4.53
psycopg2-binary==2.9.9
alembic==1.13.0
asyncpg==0.29.0
databases[postgresql]==0.8.0

# Scheduling and Background Tasks
apscheduler==3.10.4

# Additional utilities
python-dateutil==2.8.2
